Move peril to use accordion
